requirements.txt
neo4j
openai
python-dotenv
pydantic
pydantic-ai
